+++ src/input.cpp
@@ -752,12 +752,12 @@ char_event_t inputter_t::read_char(const command_handler_t &command_handler) {
                 }
                 case readline_cmd_t::func_and:
                 case readline_cmd_t::func_or: {
-                    // If previous function has right status, we keep reading tokens
+                    // If previous function has correct status, we keep reading tokens
                     if (evt.get_readline() == readline_cmd_t::func_and) {
-                        if (function_status_) return readch();
+                        // Don't return immediately, we might need to handle it here - like self-insert.
+                        if (function_status_) continue;
                     } else {
-                        assert(evt.get_readline() == readline_cmd_t::func_or);
-                        if (!function_status_) return readch();
+                        if (!function_status_) continue;
                     }
                     // Else we flush remaining tokens
                     do {